feat: add configure module to split long code
